+++ R/web_fonts.R
@@ -4,9 +4,10 @@
 #' looking for a font, in order to see examples etc, `search_web_fonts()`
 #' provide a quick lookup based on family name in the repositories supported by
 #' systemfonts (currently [Google Fonts](https://fonts.google.com) and
-#' [Font Squirrel](https://www.fontsquirrel.com)). The lookup is based on fuzzy
-#' matching provided by [utils::adist()] and the matching parameters can be
-#' controlled through `...`
+#' [Font Squirrel](https://www.fontsquirrel.com) - [Bunny Fonts](https://fonts.bunny.net/)
+#' provide the same fonts as Google Fonts but doesn't have a search API). The
+#' lookup is based on fuzzy matching provided by [utils::adist()] and the
+#' matching parameters can be controlled through `...`
 #'
 #' @param family The font family name to look for
 #' @param n_max The maximum number of matches to return
@@ -28,7 +29,7 @@ search_web_fonts <- function(family, n_max = 10, ...) {
   all <- data.frame(
     family = c(gf, fs),
     repository = c(rep(
-      c("Google Fonts", "Font Squirrel"),
+      c("Google Fonts/Bunny Fonts", "Font Squirrel"),
       c(length(gf), length(fs))
     ))
   )
@@ -231,14 +232,14 @@ require_font <- function(
 #' can be achieved through the use of stylesheets that can either be added with
 #' a `<link>` tag or inserted with an `@import` statement. This function
 #' facilitates the creation of either of these (or the bare URL to the
-#' stylesheet). It can rely on the Google Fonts or Font Library repository for
-#' serving the fonts. If the requested font is not found it can optionally hard
-#' code the data into the stylesheet.
+#' stylesheet). It can rely on the Bunny Fonts, Google Fonts and/or Font Library
+#' repositories for serving the fonts. If the requested font is not found it can
+#' optionally hard code the data into the stylesheet.
 #'
 #' @inheritParams match_fonts
 #' @param ... Additional arguments passed on to the specific functions for the
 #' repositories. Currently:
-#' * **Google Fonts:**
+#' * **Google Fonts and Bunny Fonts:**
 #'   - `text` A piece of text containing the glyphs required. Using this can
 #'     severely cut down on the size of the required download
 #'   - `display` One of `"auto"`, `"block"`, `"swap"`, `"fallback"`, or
@@ -252,8 +253,9 @@ require_font <- function(
 #' repositories be embedded as data-URLs. This is only possible if the font is
 #' available locally and in a `woff2`, `woff`, `otf`, or `ttf` file.
 #' @param repositories The repositories to try looking for the font. Currently
-#' `"Google Fonts"` and `"Font Library"` are supported. Set this to `NULL`
-#' together with `may_embed = TRUE` to force embedding of the font data.
+#' `"Bunny Fonts"`, `"Google Fonts"`, and `"Font Library"` are supported. Set
+#' this to `NULL` together with `may_embed = TRUE` to force embedding of the
+#' font data.
 #'
 #' @return A character vector with stylesheet specifications according to `type`
 #' @export
@@ -266,7 +268,7 @@ fonts_as_import <- function(
   ...,
   type = c("url", "import", "link"),
   may_embed = TRUE,
-  repositories = c("Google Fonts", "Font Library")
+  repositories = c("Bunny Fonts", "Google Fonts", "Font Library")
 ) {
   import <- character(0)
   type <- match.arg(type)
@@ -275,6 +277,13 @@ fonts_as_import <- function(
   for (repo in repositories) {
     fonts <- switch(
       tolower(repo),
+      "bunny fonts" = import_from_bunny_fonts(
+        family,
+        italic = italic,
+        weight = weight,
+        width = width,
+        ...
+      ),
       "google fonts" = import_from_google_fonts(
         family,
         italic = italic,
@@ -345,7 +354,7 @@ import_from_google_fonts <- function(
     fam <- paste0("family=", gsub(" ", "+", family[i[1]]))
 
     spec <- list()
-    spec$ital <- if (!is.null(italic)) unique(range(italic[i]))
+    spec$ital <- if (!is.null(italic)) as.character(unique(range(italic[i])))
     if (isTRUE(spec$ital == 0L)) spec$ital <- NULL
     spec$wdth <- if (!is.null(width))
       paste0(unique(range(width[i])), collapse = "..")
@@ -353,8 +362,18 @@ import_from_google_fonts <- function(
       paste0(unique(range(weight[i])), collapse = "..")
     spec <- spec[lengths(spec) != 0]
     if (length(spec) != 0) {
-      spec$sep <- ","
-      val <- paste0(do.call(paste0, spec), collapse = ";")
+      n_specs <- max(lengths(spec))
+      spec <- lapply(spec, rep_len, n_specs)
+      val <- paste0(
+        vapply(
+          seq_len(n_specs),
+          function(i) {
+            paste0(vapply(spec, `[[`, character(1), i), collapse = ",")
+          },
+          character(1)
+        ),
+        collapse = ";"
+      )
       spec <- paste0(names(spec), collapse = ",")
       fam <- paste0(fam, ":", spec, "@", val)
     }
@@ -384,6 +403,38 @@ import_from_google_fonts <- function(
     missing <- seq_along(family)
   }
   structure(url %||% character(), no_match = missing)
+}
+import_from_bunny_fonts <- function(
+  family,
+  italic = NULL,
+  weight = NULL,
+  width = NULL,
+  ...,
+  text = NULL,
+  display = "swap"
+) {
+  url <- import_from_google_fonts(
+    family = family,
+    italic = italic,
+    weight = weight,
+    width = width,
+    ...,
+    text = text,
+    display = display
+  )
+  success <- try(suppressWarnings(readLines(url, n = 1)), silent = TRUE)
+  if (inherits(success, "try-error") || any(grepl("Error: API Error", success))) {
+    return(structure(character(), no_match = seq_along(family)))
+  }
+  structure(
+    sub(
+      "https://fonts.googleapis.com/css2?",
+      "https://fonts.bunny.net/css2?",
+      url,
+      fixed = TRUE
+    ),
+    no_match = attr(url, "no_match")
+  )
 }
 import_from_font_library <- function(family, ...) {
   family <- gsub(" ", "-", tolower(family))